moved generic error
